Add new ManualOpMode structure
