+++ build.dependencies.gradle
@@ -12,6 +12,6 @@ dependencies {
     implementation 'org.firstinspires.ftc:Hardware:10.0.0'
     implementation 'org.firstinspires.ftc:FtcCommon:10.0.0'
     implementation 'org.firstinspires.ftc:Vision:10.0.0'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.appcompat:appcompat:1.7.0'
 }
